requirements.txt
streamlit
tensorflow
numpy
Pillow
python-dotenv
ollama==0.3.0
langchain==0.2.10
langchain-community==0.2.9
langchain-experimental==0.0.62
langchain-ollama==0.1.0
pandas==2.2.2
openpyxl==3.1.5
tabulate==0.9.0